Move card header component to common
