+++ js/components/developer/create-job-screen/timeCard.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Card, CardItem } from 'native-base';
-import CardHeader from './cardHeader';
+import CardHeader from '../common/cardHeader';
 import RadioWithNote from './radioWithNote';
 
 // Temporary data, will be moved and handled in another way in the future.
